Copy captured scalars into scope, remove rewriting
With this approach we copy the value of scalar variables, that get mapped
from the host to the target, into a local copy of the variable
declaration. Hopefully, this way we can get around the limitations of
OpenMP's scalar handling (in clang).
This copies every scalar regardless if it is used in an OpenMP
construct.